project_3/amazon-mockup-e-commerce/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
passlib[bcrypt]
streamlit
requests
pydantic[email]
python-dotenv